requirements.txt
python-dotenv
langchain_core
langchain
langchain_google_genai
google-generativeai
retry
transformers
tqdm
gensim
numpy
nltk
unidecode
deep_translator
pdfminer.six
torch
scikit-learn
requests_html
fake_useragent
datetime
beautifulsoup4
configparser
pandas
opencv-python
camelot-py
faiss-cpu
sentence-transformers
ghostscript